Add test createToDo
